Ändere den Kalorien-Typ in der FastFood-Klasse von int auf double für genauere Werte.

+++ klausuraufgaben/src/main/java/com/merzlab/Shop/FastFood.java
@@ -3,7 +3,7 @@ package com.merzlab.Shop;
 public record FastFood(
         String name,
         FastFoodCategory category,
-        int calories,
+        double calories,
         boolean vegetarian
 ) implements Comparable<FastFood> {
